[TrainingList] add empty list placeholder and confirm uppon deletion

+++ app/src/main/java/com/wile/app/ui/main/TrainingListingFragment.kt
@@ -1,9 +1,12 @@
 package com.wile.app.ui.main
 
+import android.content.Context
+import android.content.DialogInterface
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.appcompat.app.AlertDialog
 import androidx.fragment.app.viewModels
 import com.wile.app.R
 import com.wile.core.databinding.DataBindingFragment
@@ -11,6 +14,7 @@ import com.wile.app.databinding.FragmentTrainingListBinding
 import com.wile.core.extensions.exhaustive
 import com.wile.app.ui.adapter.TrainingAdapter
 import com.wile.app.ui.add.AddActivity
+import com.wile.app.ui.add.QuickAddActivity
 import com.wile.app.ui.add.TabataAddActivity
 import com.wile.database.model.Training
 import com.wile.database.model.TrainingTypes
@@ -20,6 +24,7 @@ import dagger.hilt.android.AndroidEntryPoint
 class TrainingListingFragment : DataBindingFragment() {
 
     private val viewModel: TrainingListingViewModel by viewModels()
+    private var workoutId = -1
     private val binding: FragmentTrainingListBinding by binding(R.layout.fragment_training_list)
     private val adapter by lazy { TrainingAdapter(
         onDeleteTraining = ::onDeleteTraining,
@@ -32,15 +37,44 @@ class TrainingListingFragment : DataBindingFragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
+
+        workoutId = arguments?.getInt(WORKOUT_ID, -1) ?: kotlin.run { -1 }
+
         binding.apply {
             adapter = this@TrainingListingFragment.adapter
             viewModel = this@TrainingListingFragment.viewModel
         }
+
         return binding.root
     }
 
+    override fun onAttach(context: Context) {
+        binding.placeholderAddTrainingBtn.setOnClickListener {
+            if (workoutId>=0) {
+                startActivity(QuickAddActivity.newIntent(context, workoutId))
+            }
+        }
+        super.onAttach(context)
+    }
+
     private fun onDeleteTraining(training: Training) {
-        viewModel.deleteTraining(training.id)
+        if (adapter.itemCount == 1){
+            activity?.let {
+                val builder = AlertDialog.Builder(it)
+                builder.setMessage(R.string.delete_last_training)
+                    .setPositiveButton(R.string.delete
+                    ) { _, _ ->
+                        viewModel.deleteTraining(training.id)
+                    }
+                    .setNegativeButton(R.string.cancel
+                    ) { dialog, _ ->
+                        dialog.cancel()
+                    }
+                builder.create().show()
+            } ?: kotlin.run { viewModel.deleteTraining(training.id) }
+        } else {
+            viewModel.deleteTraining(training.id)
+        }
     }
 
     private fun onMoveTraining(trainings: List<Training>) {
